requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart==0.0.5
requests==2.27.1
pytest==6.2.5
python-dotenv==0.19.2
joblib==1.2.0
scikit-learn==1.2.2
pandas>=2.0.2
azureml-sdk
scipy==1.7.3